AD-HOC feat (*): Initial commit
With the release of the 2.2.3 Magento release, the behaviour of
symlinking in stateful directories such as media and log appears to have
broken. Notably, this is visible in media where when a user tries to
upload an image they are rejected with an opaque notice that "media is
not in the application path".

Symlinks are a common mechanism to help implement zero down time
deployment. Other mechanisms are also possible such as bind mounts or
hard links, but symlinks present some unique characteristics that are
desirable with release management (such as being and safely easily
removable)

The security implications of this change are not yet clear. While the
author is confident that in the environment this fix intends to be
deployed this will be safe solution, consideration should be given to
whether symlinks are still safe in other situations as they're
essentially expanding the directory tree of the application.

== Stakeholder Impact ==

=== Project Owner ====

This allows administrators to upload image again.

=== Developers ===

Packaging the application as a separate module allows easy packaging and
distribution to other projects, as well as removal when the fix
invariably does land.

== Community ==

This issue is super common, and packaging an extension seems like a nice
thing to do to make it easily fixable, as well as centrally track
whether this turns out to be a horrible idea.

== Design Notes ==

=== Coding Standard ===

This follows the Sitewards standards, an extension of PSR-2 which
notably includes Hungarian Notation.